feat: split tests for interactive-wrapper
